+++ server/src/walking.cpp
@@ -724,7 +724,7 @@ bool cMovement::consumeStamina( P_PLAYER pChar, bool running ) {
 	}
 
 	// Calculate the stones we weight too much
-	int overWeight = pChar->weight() - pChar->maxWeight();
+	int overWeight = ( int )( pChar->weight() - pChar->maxWeight() );
 	bool mounted = pChar->atLayer(cBaseChar::Mount) != 0;
 	bool update = false;
